pyproject.toml
[tool.poetry]
name = "bnn-opf"
version = "0.1.0"
description = "Bayesian Neural Networks for the OPF problem"
authors = ["kaarthiksundar <kaarthik01sundar@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "bnn_opf"}]

[tool.poetry.dependencies]
python = "^3.11,<3.13"
typer = "^0.12.3"
logging = "^0.4.9.6"
numpy = "^1.26.0"
scipy = "^1.14.0"
gridx-egret = {path = "Egret"}
torch_geometric = {git = "https://github.com/pyg-team/pytorch_geometric.git", branch="master"}
jax = "^0.4.31"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
